Move propagator-specific functions to module
