+++ src/poliastro/core/angles.py
@@ -24,87 +24,10 @@ def _kepler_equation_prime_hyper(F, M, ecc):
 
 
 @jit
-def _kepler_equation_parabolic(D, M, ecc):
-    return D_to_M_near_parabolic(ecc, D) - M
-
-
-@jit
-def _kepler_equation_prime_parabolic(D, M, ecc):
-    x = (ecc - 1.0) / (ecc + 1.0) * (D ** 2)
-    assert abs(x) < 1
-    S = dS_x_alt(ecc, x)
-    return np.sqrt(2.0 / (1.0 + ecc)) + np.sqrt(2.0 / (1.0 + ecc) ** 3) * (D ** 2) * S
-
-
-@jit
-def S_x(ecc, x, atol=1e-12, maxiter=100):
-    S = 0
-    k = 0
-    while k < maxiter:
-        S_old = S
-        S += (ecc - 1 / (2 * k + 3)) * x ** k
-        k += 1
-        if abs(S - S_old) < atol:
-            return S
-    else:
-        raise RuntimeError("Function did not converge")
-
-
-@jit
-def dS_x_alt(ecc, x, atol=1e-12, maxiter=100):
-    # Notice that this is not exactly
-    # the partial derivative of S with respect to D,
-    # but the result of arranging the terms
-    # in section 4.2 of Farnocchia et al. 2013
-    S = 0
-    k = 0
-    while k < maxiter:
-        S_old = S
-        S += (ecc - 1 / (2 * k + 3)) * (2 * k + 3) * x ** k
-        k += 1
-        if abs(S - S_old) < atol:
-            return S
-    else:
-        raise RuntimeError("Function did not converge")
-
-
-@jit
-def d2S_x_alt(ecc, x, atol=1e-12, maxiter=100):
-    # Notice that this is not exactly
-    # the second partial derivative of S with respect to D,
-    # but the result of arranging the terms
-    # in section 4.2 of Farnocchia et al. 2013
-    # Also, notice that we are not using this function yet
-    S = 0
-    k = 0
-    while k < maxiter:
-        S_old = S
-        S += (ecc - 1 / (2 * k + 3)) * (2 * k + 3) * (2 * k + 2) * x ** k
-        k += 1
-        if abs(S - S_old) < atol:
-            return S
-    else:
-        raise RuntimeError("Function did not converge")
-
-
-@jit
-def D_to_M_near_parabolic(ecc, D):
-    x = (ecc - 1.0) / (ecc + 1.0) * (D ** 2)
-    assert abs(x) < 1
-    S = S_x(ecc, x)
-    return (
-        np.sqrt(2.0 / (1.0 + ecc)) * D + np.sqrt(2.0 / (1.0 + ecc) ** 3) * (D ** 3) * S
-    )
-
-
-@jit
 def newton(regime, x0, args=(), tol=1.48e-08, maxiter=50):
     p0 = 1.0 * x0
     for iter in range(maxiter):
-        if regime == "parabolic":
-            fval = _kepler_equation_parabolic(p0, *args)
-            fder = _kepler_equation_prime_parabolic(p0, *args)
-        elif regime == "hyperbolic":
+        if regime == "hyperbolic":
             fval = _kepler_equation_hyper(p0, *args)
             fder = _kepler_equation_prime_hyper(p0, *args)
         else:
@@ -116,6 +39,7 @@ def newton(regime, x0, args=(), tol=1.48e-08, maxiter=50):
         if abs(p - p0) < tol:
             return p
         p0 = p
+
     return 1.0
 
 
@@ -356,28 +280,6 @@ def M_to_D(M):
 
 
 @jit
-def M_to_D_near_parabolic(M, ecc):
-    """Parabolic eccentric anomaly from mean anomaly, near parabolic case.
-
-    Parameters
-    ----------
-    M : float
-        Mean anomaly in radians.
-    ecc : float
-        Eccentricity (~1).
-
-    Returns
-    -------
-    D : float
-        Parabolic eccentric anomaly.
-
-    """
-    D0 = M_to_D(M)
-    D = newton("parabolic", D0, args=(M, ecc), maxiter=100)
-    return D
-
-
-@jit
 def E_to_M(E, ecc):
     """Mean anomaly from eccentric anomaly.
 
@@ -437,103 +339,6 @@ def D_to_M(D):
 
     """
     M = D + D ** 3 / 3
-    return M
-
-
-@jit
-def D_to_M_near_parabolic_alt(D, ecc):
-    """Mean anomaly from eccentric anomaly in the near parabolic region.
-
-    Parameters
-    ----------
-    D : float
-        Parabolic eccentric anomaly.
-    ecc : float
-        Eccentricity.
-
-    Returns
-    -------
-    M : float
-        Mean anomaly.
-
-    """
-    M = _kepler_equation_parabolic(D, 0.0, ecc)
-    return M
-
-
-@jit
-def M_to_nu(M, ecc, delta=1e-2):
-    """True anomaly from mean anomaly.
-
-    .. versionadded:: 0.4.0
-
-    Parameters
-    ----------
-    M : float
-        Mean anomaly in radians.
-    ecc : float
-        Eccentricity.
-    delta : float (optional)
-        threshold of near-parabolic regime definition (from Davide Farnocchia et al)
-
-    Returns
-    -------
-    nu : float
-        True anomaly.
-
-    Examples
-    --------
-    >>> from numpy import radians, degrees
-    >>> degrees(M_to_nu(radians(30.0), 0.06))
-    33.67328493021166
-
-    """
-    if ecc > 1 + delta:
-        F = M_to_F(M, ecc)
-        nu = F_to_nu(F, ecc)
-    elif ecc < 1 - delta:
-        E = M_to_E(M, ecc)
-        nu = E_to_nu(E, ecc)
-    else:
-        D = M_to_D_near_parabolic(M, ecc)
-        nu = D_to_nu(D)
-    return nu
-
-
-@jit
-def nu_to_M(nu, ecc, delta=1e-2):
-    """Mean anomaly from true anomaly.
-
-    .. versionadded:: 0.4.0
-
-    Parameters
-    ----------
-    nu : float
-        True anomaly in radians.
-    ecc : float
-        Eccentricity.
-    delta : float (optional)
-        Threshold of near-parabolic regime definition (from Davide Farnocchia et al)
-
-    Returns
-    -------
-    M : float
-        Mean anomaly.
-
-    """
-    if ecc > 1 + delta:
-        F = nu_to_F(nu, ecc)
-        M = F_to_M(F, ecc)
-    elif ecc < 1 - delta:
-        E = nu_to_E(nu, ecc)
-        M = E_to_M(E, ecc)
-    else:
-        D = nu_to_D(nu)
-        # FIXME: This should be
-        # M = D_to_M_near_parabolic(D, ecc)
-        # but the discrimination of near-parabolic regions here
-        # is not correct and the function might not converge
-        M = D_to_M_near_parabolic_alt(D, ecc)
     return M
