+++ tests/tests_twobody/test_angles.py
@@ -10,7 +10,7 @@ from poliastro.core.elements import coe2mee, coe2rv, mee2coe, rv2coe
 from poliastro.twobody import angles
 
 # Data from Schlesinger & Udick, 1912
-ANGLES_DATA = [
+ELLIPTIC_ANGLES_DATA = [
     # ecc, M (deg), nu (deg)
     (0.0, 0.0, 0.0),
     (0.05, 10.0, 11.06),
@@ -122,25 +122,25 @@ def test_true_to_eccentric_hyperbolic():
 
 
 def test_mean_to_true():
-    for row in ANGLES_DATA:
+    for row in ELLIPTIC_ANGLES_DATA:
         ecc, M, expected_nu = row
         ecc = ecc * u.one
         M = M * u.deg
         expected_nu = expected_nu * u.deg
 
-        nu = angles.M_to_nu(M, ecc)
+        nu = angles.E_to_nu(angles.M_to_E(M, ecc), ecc)
 
         assert_quantity_allclose(nu, expected_nu, rtol=1e-4)
 
 
 def test_true_to_mean():
-    for row in ANGLES_DATA:
+    for row in ELLIPTIC_ANGLES_DATA:
         ecc, expected_M, nu = row
         ecc = ecc * u.one
         expected_M = expected_M * u.deg
         nu = nu * u.deg
 
-        M = angles.nu_to_M(nu, ecc)
+        M = angles.E_to_M(angles.nu_to_E(nu, ecc), ecc)
 
         assert_quantity_allclose(M, expected_M, rtol=1e-4)
 
@@ -152,7 +152,7 @@ def test_true_to_mean_hyperbolic():
     ecc = 2.7696 * u.one
     expected_M = 11.279 * u.rad
 
-    M = angles.nu_to_M(nu, ecc)
+    M = angles.F_to_M(angles.nu_to_F(nu, ecc), ecc)
 
     assert_quantity_allclose(M, expected_M, rtol=1e-4)
 
@@ -164,7 +164,7 @@ def test_mean_to_true_hyperbolic():
     ecc = 2.7696 * u.one
     expected_nu = 100 * u.deg
 
-    nu = angles.M_to_nu(M, ecc)
+    nu = angles.F_to_nu(angles.M_to_F(M, ecc), ecc)
 
     assert_quantity_allclose(nu, expected_nu, rtol=1e-4)
 
@@ -185,8 +185,8 @@ def test_flight_path_angle():
 )
 @pytest.mark.parametrize("ecc", [3200 * u.one, 1.5 * u.one])
 def test_mean_to_true_hyperbolic_highecc(expected_nu, ecc):
-    M = angles.nu_to_M(expected_nu, ecc)
-    nu = angles.M_to_nu(M, ecc)
+    M = angles.F_to_M(angles.nu_to_F(expected_nu, ecc), ecc)
+    nu = angles.F_to_nu(angles.M_to_F(M, ecc), ecc)
     assert_quantity_allclose(nu, expected_nu, rtol=1e-4)
